% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spongebobysay.R
\docType{data}
\name{ascii_spongebob}
\alias{ascii_spongebob}
\title{ASCII Spongebob}
\format{An object of class \code{character} of length 1.}
\usage{
ascii_spongebob
}
\description{
ASCII Mocking Spongebob used by \code{\link{spongebobsay}}.
}
\examples{
cat(ascii_spongebob)
}
\keyword{datasets}
