% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imdb.R
\docType{data}
\name{imdb}
\alias{imdb}
\title{IMDB data for SpongeBob SquarePants}
\format{
A tibble with 53351 observations of 27 variables.
\describe{
  \item{rating}{Mean rating, out of 10}
  \item{votes}{Number of votes the episode received}
  \item{episode}{Season and episode within season formatted S*, E*}
}
}
\source{
\url{https://www.imdb.com/title/tt0206512/}
}
\usage{
imdb
}
\description{
Data scraped from imdb.com on October 16, 2022. Since SpongeBob is an ongoing series, these ratings will change over time.
}
\details{
Join with the sbob data set by using dplyr::left_join(sbob, imdb, by = c("imdb_episode" = "episode"))
}
\keyword{datasets}
