\name{ravioliravioli}
\alias{ravioliravioli}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Test
}
\description{
Test
}
\usage{
ravioliravioli(n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{
number 
}
}
\details{
Something
}
\value{
greg
}
\references{
none
}
\author{
S.Bob
}
\note{
None
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
test 
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (n) 
{
    out <- rnorm(n)
    return(out)
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
