\name{spongebob-package}
\alias{spongebob-package}
\alias{spongebob}
\docType{package}
\title{
\packageTitle{spongebob}
}
\description{
\packageDescription{spongebob}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{spongebob}
\packageIndices{spongebob}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{spongebob}

Maintainer: \packageMaintainer{spongebob}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
