% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spongebobysay.R
\name{spongebobsay}
\alias{spongebobsay}
\alias{spongebobthink}
\alias{spongebobwhisper}
\title{Spongebob ASCII Messages : SPongEBob aScii MeSsaGeS}
\usage{
spongebobsay(what)

spongebobthink(what)

spongebobwhisper(what)
}
\arguments{
\item{what}{a length-1 character string. A leNGTh-1 chARActeR StRInG.}
}
\description{
A set of functions that generate ASCII art with speech bubbles
in the style of \href{https://en.wikipedia.org/wiki/Cowsay}{cowsay}.

a SeT Of fUNcTiONS tHat gEnerATE asCiI arT wiTH SpEEcH BUbBles In thE
StylE Of \href{https://en.wikipedia.org/wiki/Cowsay}{cOWsaY}.
}
\examples{
speech <- spongebobsay(
    "globalVariables is a hideous hack and I will never use it."
)
cat(speech)

thought <- spongebobthink("@hadley you shouldn't say you'll never use things when only two years later you think it's fine -- hadley")
cat(thought)

whisper <- spongebobwhisper("it's still a hideous hack, but you just have to suck it up if you want your package on CRAN")
cat(whisper)

}
