% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tospongebob.R
\name{tospongebob.default}
\alias{tospongebob.default}
\title{SpongeBob-Case Conversion for List-like Objects}
\usage{
\method{tospongebob}{default}(x, ..., convert.names = TRUE)
}
\arguments{
\item{x}{object with text to be converted to Mocking SpongeBob case.}

\item{...}{onlY hERe tO sATisFy R CMd CHeCk.}

\item{convert.names}{logical, indicating whether or not to convert the names
of the object}
}
\value{
object with its text converted to Mocking SpongeBob case. ObjECt WiTH
iTS tEXt CONvErTeD TO mOCkinG SponGEBob CasE.
}
\description{
Convert list-like objects to
\href{https://knowyourmeme.com/memes/mocking-spongebob}{Mocking SpongeBob
case}. This is the default S3 method the S3 generic
\code{\link{tospongebob}}. Generally, you will not need to use this method
directly; instead, use \code{\link{tospongebob}}. If you really need to use
this method, you will need to do \code{spongebob:::tospongebob.default}
because it is an internal method.
}
\examples{
nicktoons <- list(
    spongebob = list(
        episodes = c("Help Wanted", "Reef Blower", "Tea at the Treedome"),
        characters = c("SpongeBob", "Patrick", "Squidward", "Sandy")
    ),
    hey_arnold = list(
        episodes = c("Downtown as Fruits", "Eugene's Bike"),
        characters = c("Arnold", "Gerald", "Helga")
    )
)
spongebob:::tospongebob.default(nicktoons)
}
\seealso{
\code{\link{tospongebob}}
}
