% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spongebob.R
\name{tospongebob}
\alias{tospongebob}
\title{Spongebob-Case Conversion : sPOngEBOb-CasE cOnVErsIoN}
\usage{
tospongebob(x)
}
\arguments{
\item{x}{a character vector. A CHarACTer vEcToR.}
}
\value{
A character vector of the same length as \code{x}.

a ChaRaCteR vecTor OF The SAme LenGth As \code{x}.
}
\description{
Translate character vectors to spongebob case.

trAnSLatE ChaRacTeR veCtOrS To SponGebOB CaSE
}
\examples{
tospongebob("Spongebob-Case Conversion")
tospongebob(x = c(
    "Translate character vectors to spongebob case.",
    "a character vector.",
    "A character vector of the same length as"
))
}
