% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spongebobsay.R
\name{ascii_spongebob}
\alias{ascii_spongebob}
\title{ASCII Spongebob}
\usage{
ascii_spongebob(print = TRUE)
}
\arguments{
\item{print}{a length-1 logical flag for whether to print output to console.
A LenGTH-1 lOGicaL flaG foR wHethEr tO prInT ResUlt TO cONsolE.}
}
\value{
a character string containing the ASCII art. Use
\code{\link[base]{cat}} to print with proper formatting.

A CHaRACtER stRINg cONtaINiNG ThE ASciI ARt. uSe \code{\link[base]{cat}} TO
prInT WiTH prOper FORmatTiNg.
}
\description{
ASCII Mocking Spongebob used by \code{\link{spongebobsay}}.
}
\examples{
ascii_spongebob()

art <- ascii_spongebob(print = FALSE)
cat(art)
}
