% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spongebobsay.R
\name{spongebobsay}
\alias{spongebobsay}
\alias{spongebobthink}
\alias{spongebobwhisper}
\title{SpongeBob ASCII Messages : sPongEBoB aScii MesSAgEs}
\usage{
spongebobsay(what, print = TRUE)

spongebobthink(what, print = TRUE)

spongebobwhisper(what, print = TRUE)
}
\arguments{
\item{what}{a length-1 character string. A leNGTh-1 chARActeR StRInG.}

\item{print}{a length-1 logical flag for whether to print output to console.
A LenGTH-1 lOGicaL flaG foR wHethEr tO prInT ResUlt TO cONsolE.}
}
\value{
a character string containing the ASCII art. Use
\code{\link[base]{cat}} to print with proper formatting.

A CHaRACtER stRINg cONtaINiNG ThE ASciI ARt. uSe \code{\link[base]{cat}} TO
prInT WiTH prOper FORmatTiNg.
}
\description{
A set of functions that generate SpongeBob ASCII art with speech
bubbles in the style of \href{https://en.wikipedia.org/wiki/Cowsay}{cowsay}.

a seT OF fuNctiOnS tHat GEnerAte sPONgeBOb AScIi aRT wiTH SpeEcH buBBLeS In
THe sTylE OF \href{https://en.wikipedia.org/wiki/Cowsay}{cOWsaY}.
}
\examples{
spongebobsay("globalVariables is a hideous hack and I will never use it.")

spongebobthink(
    "@hadley you shouldn't say you'll never use things when only two years
    later you think it's fine -- hadley"
)

spongebobwhisper(
    "it's still a hideous hack, but you just have to suck it up if you want
    your package on CRAN"
)

}
