% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tospongebob.R
\name{tospongebob}
\alias{tospongebob}
\alias{tospongebob.array}
\alias{tospongebob.factor}
\alias{tospongebob.fortune}
\alias{tospongebob.environment}
\alias{tospongebob.function}
\alias{tospongebob.ggplot}
\alias{tospongebob.matrix}
\alias{tospongebob.table}
\title{SpongeBob-Case Conversion : sPOngEBOb-CasE cOnVErsIoN}
\usage{
tospongebob(x, ...)
}
\arguments{
\item{x}{object with text to be converted to Mocking SpongeBob case.}

\item{...}{onlY hERe tO sATisFy R CMd CHeCk.}
}
\value{
object with its text converted to Mocking SpongeBob case. ObjECt WiTH
iTS tEXt CONvErTeD TO mOCkinG SponGEBob CasE.
}
\description{
\code{tospongebob} is an S3 generic function for converting text
in R objects to \href{https://knowyourmeme.com/memes/mocking-spongebob}{
Mocking SpongeBob case}. The core method for character vectors will return
the input character vector with case-converted elements. For other objects,
it will attempt to appropriately find text and convert them.

\code{tospongebob} IS an S3 gENErIC FunCtIoN fOR cOnveRtIng Text IN r obJEcts
To \href{https://knowyourmeme.com/memes/mocking-spongebob}{MocKinG SpOnGEbob
CasE}. tHe cOrE MEtHOd FoR CharActEr VECtorS wIll RetURN tHe INPuT chARaCTer
VEcTOr WItH cASe-cONVeRTed ELeMenTS. foR oThER objECTs, iT wIlL aTtemPT tO
APprOPrIAtELy fInd tEXt aNd cONvERT tHEm.
}
\details{
This section describes the (non-exported) S3 methods in more detail.

The core \bold{\code{\link{tospongebob.character}}} method takes an input character
vector and returns the same vector with its elements converted to Mocking
SpongeBob case. Names of the vector are also converted unless parameter
\code{convert.names = FALSE}. This method is then leveraged by all other
methods.

The default \bold{\code{\link{tospongebob.default}}} method will iterate through any
elements of list-like objects and attempt to convert those. It will also
convert the names of the object unless parameter \code{convert.names = FALSE}.
NULL values are passed through as NULL.

Additional S3 methods for the following classes have been implemented:
\describe{
    \item{\bold{\code{tospongebob.array}}}{: convert data and dimnames of an
    array}
    \item{\bold{\code{\link{tospongebob.data.frame}}}}{: convert row names,
    column names, and convertible columns of a data.frame}
    \item{\bold{\code{tospongebob.environment}}}{: create a copy of an
    environment with its objects converted}
    \item{\bold{\code{tospongebob.factor}}}{: convert levels of factor
    vectors}
    \item{\bold{\code{tospongebob.fortune}}}{: convert the content of a
    \link[fortunes:fortunes]{fortunes::fortune} object}
    \item{\bold{\code{tospongebob.function}}}{: deparse function into a
    character vector, convert, and then combine into one string block}
    \item{\bold{\code{tospongebob.ggplot}}}{: convert text labels in
    \code{\link[ggplot2:ggplot]{ggplot2::ggplot}} object}
    \item{\bold{\code{matrix}}}{: convert data and dimnames of a
    matrix}
    \item{\bold{\code{tospongebob.table}}}{: convert data and dimnames of a
    table}
}
}
\examples{
tospongebob("SpongeBob-Case Conversion")
tospongebob(c("S3 generic function", "Mocking SpongeBob case"))

tospongebob(packageDescription("base"))

df <- head(datasets::CO2)
tospongebob(df)

tospongebob(summary(df))

nicktoons <- list(
    spongebob = list(
        episodes = c("Help Wanted", "Reef Blower", "Tea at the Treedome"),
        characters = c("SpongeBob", "Patrick", "Squidward", "Sandy")
    ),
    hey_arnold = list(
        episodes = c("Downtown as Fruits", "Eugene's Bike"),
        characters = c("Arnold", "Gerald", "Helga")
    )
)
tospongebob(nicktoons)
}
