% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spongebob.R
\name{tospongebob}
\alias{tospongebob}
\title{sPOnGEBoB-caSE cONVersIon}
\usage{
tospongebob(x)
}
\arguments{
\item{x}{a ChaRActeR vECtOr.}
}
\value{
a CharAcTeR VECtOR oF THe SamE LeNGTh AS \code{x}
}
\description{
trAnSLaTe ChARAcTers IN chARaCter veCToRS tO sPONgebOB caSe.
}
\examples{
tospongebob("spongebob-case conversion")
tospongebob(x = c(
    "translate characters in character vectors to spongebob case.",
    "a character vector.",
    "a character vector of the same length as"
))
}
