% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tospongebob.R
\name{tospongebob}
\alias{tospongebob}
\title{SpongeBob-Case Conversion : sPOngEBOb-CasE cOnVErsIoN}
\usage{
tospongebob(x)
}
\arguments{
\item{x}{a character vector. A CHarACTer vEcToR.}
}
\value{
A character vector of the same length as \code{x}.

a ChaRaCteR vecTor OF The SAme LenGth As \code{x}.
}
\description{
Translate character vectors to
\href{https://knowyourmeme.com/memes/mocking-spongebob}{Mocking SpongeBob
case}.

tRANsLate ChaRACtER vECtoRs tO
\href{https://knowyourmeme.com/memes/mocking-spongebob}{mOcKiNg spOngeBoB
caSe}.
}
\examples{
tospongebob("SpongeBob-Case Conversion")
tospongebob(x = c(
    "Translate character vectors to Mocking SpongeBob case",
    "a character vector.",
    "A character vector of the same length as"
))
}
